+++ ve/unit/test_compound_obj.py
@@ -21,6 +21,7 @@ from vsc.types import rand_uint16_t
 from vsc.attrs import rand_attr, attr
 import vsc
 from vsc_test_case import VscTestCase
+from vsc.methods import raw_mode
 
 
 class TestCompoundObj(VscTestCase):
@@ -130,14 +131,7 @@ class TestCompoundObj(VscTestCase):
 
             @vsc.constraint
             def ab_c(self):
-                with vsc.foreach(self.c1, idx=True) as idx:
-                    with vsc.if_then(idx == 0):
-                        self.c1[idx].a < 6
-#                     with vsc.if_then(idx == 0):
-#                         with vsc.if_then(idx == 0):
-#                             self.c1[idx].a == self.c2[idx].x
                 self.c1[0].a < 6
-#                self.c1[0].a == self.c2[0].x
                 pass
 
         @vsc.randobj
@@ -165,4 +159,325 @@ class TestCompoundObj(VscTestCase):
         print(inst.c1[0].a)
         print(inst.c2[1].x)
         
+    def test_two_layer_small(self):
+        
+        @vsc.randobj
+        class Top(object):
+            def __init__(self):
+                self.c1 = vsc.rand_list_t(vsc.attr(Sub1()))
+                self.c1.append(vsc.attr(Sub1()))
+                
+            @vsc.constraint
+            def array_c(self):
+                self.c1[0].c2[0].a == 2
+                
+        @vsc.randobj
+        class Sub1(object):
+            
+            def __init__(self):
+                self.c2 = vsc.rand_list_t(vsc.attr(Sub2()))
+                self.c2.append(vsc.attr(Sub2()))
+                self.c3 = vsc.rand_list_t(vsc.attr(Sub2()))
+                self.c3.append(vsc.attr(Sub2()))
+                
+        @vsc.randobj
+        class Sub2(object):
+            
+            def __init__(self):
+                self.a = vsc.rand_uint8_t()
+                
+        t = Top()
+                
+    def test_one_layer_small(self):
+        
+        @vsc.randobj
+        class Top(object):
+            def __init__(self):
+                s = Sub1()
+                si = Sub1()
+                print("--> c1.create")
+                self.c1 = vsc.rand_list_t(vsc.attr(s))
+                print("<-- c1.create")
+                self.c1.append(vsc.attr(si))
+                
+            @vsc.constraint
+            def array_c(self):
+                self.c1[0].a == 2
+                
+        @vsc.randobj
+        class Sub1(object):
+            
+            def __init__(self):
+                self.a = vsc.rand_uint8_t()
+                
+        t = Top()
+                
+    def test_two_layer(self):
+        import vsc 
+
+        @vsc.randobj
+        class Parent:
+            def __init__(self):
+                self.id = 0
+                self.c1 = vsc.rand_list_t(vsc.rand_attr(Child1()))
+                for i in range(2):    
+                    self.c1.append(vsc.rand_attr(Child1()))      
+
+                self.c2 = vsc.rand_list_t(vsc.rand_attr(Child2()))
+                for i in range(2):
+                    self.c2.append(vsc.rand_attr(Child2()))
+
+            @vsc.constraint
+            def parent_c(self):
+                self.c1[0].a[1].value == self.c2[0].x[1].value       # Multi-level 
+                pass
+#                self.c1[0].a[1].value == 2
+
+#                with vsc.foreach(self.c1, idx=True) as i:
+#                    self.c1[i].a[0].value == self.c2[i].x[0].value
+
+
+        @vsc.randobj
+        class Field():
+            def __init__(self, name, def_value):
+                self.name = name
+                self.value = vsc.rand_uint8_t(def_value)
+
+        @vsc.randobj
+        class Child1:
+            def __init__(self):
+                self.a = vsc.rand_list_t(vsc.rand_attr(Field('an', 10)))
+                for i in range(2):    
+                    self.a.append(vsc.rand_attr(Field('an', 10)))
+
+                self.b = vsc.rand_list_t(vsc.rand_attr(Field('bn', 10)))
+                for i in range(1):    
+                    self.b.append(vsc.rand_attr(Field('bn', 10)))
+    
+            @vsc.constraint
+            def test_c(self):
+#                self.a[0].value < self.a[1].value
+#                self.a[0].value == self.a[1].value
+                pass
+
+        @vsc.randobj
+        class Child2:
+            def __init__(self):
+                self.x = vsc.rand_list_t(vsc.rand_attr(Field('x', 10)))
+                for i in range(2):
+                    self.x.append(vsc.rand_attr(Field('x', 10)))
+
+                self.y = vsc.rand_list_t(vsc.rand_attr(Field('y', 10)))
+                for i in range(1):    
+                    self.y.append(vsc.rand_attr(Field('y', 10)))
+    
+            @vsc.constraint
+            def test_c(self):
+                self.x[0].value < self.x[1].value
+                pass
+
+        inst=Parent()
+        inst.randomize(debug=0)
+       
+        for i in range(inst.c2.size):
+            print("c[%d].x[0].value=%d c2[%d].x[1].value=%d" % (
+                i, inst.c2[i].x[0].value, i, inst.c2[i].x[1].value))
+            with vsc.raw_mode():
+                print("%s %s" % (
+                    inst.c2[i].x[0].value.get_model().fullname, 
+                    inst.c2[i].x[1].value.get_model().fullname))
+        
+        for i in range(inst.c2.size):
+#            self.assertEqual(inst.c1[i].a[0].value, inst.c1[i].a[1].value)
+            self.assertLess(inst.c2[i].x[0].value, inst.c2[i].x[1].value)
+            
+        print(inst.c1[0].a[0].value)
+        print(inst.c1[0].a[1].value)
+        print(inst.c2[0].x[0].value)        
+
+    def test_two_layer_1(self):
+        import vsc 
+
+        @vsc.randobj
+        class Parent:
+            def __init__(self):
+                self.id = 0
+                self.c1 = vsc.rand_list_t(vsc.rand_attr(Child1()))
+                for i in range(10):    
+                    self.c1.append(vsc.rand_attr(Child1()))      
+
+                self.c2 = vsc.rand_list_t(vsc.rand_attr(Child2()))
+                for i in range(10):
+                    self.c2.append(vsc.rand_attr(Child2()))
+
+            @vsc.constraint
+            def parent_c(self):
+                pass
+                self.c1[0].a[1].value == self.c2[0].x[1].value       # Multi-level 
+
+#                with vsc.foreach(self.c1, idx=True) as i:
+#                    self.c1[i].a[0].value == self.c2[i].x[0].value
+
+
+        @vsc.randobj
+        class Field():
+            def __init__(self, name, def_value):
+                self.name = name
+                self.value = vsc.rand_uint8_t(def_value)
+
+        @vsc.randobj
+        class Child1:
+            def __init__(self):
+                self.a = vsc.rand_list_t(vsc.rand_attr(Field('an', 10)))
+                for i in range(5):    
+                    self.a.append(vsc.rand_attr(Field('an', 10)))
+
+                self.b = vsc.rand_list_t(vsc.rand_attr(Field('bn', 10)))
+                for i in range(5):    
+                    self.b.append(vsc.rand_attr(Field('bn', 10)))
+    
+            @vsc.constraint
+            def test_c(self):
+#                self.a[0].value < self.a[1].value
+#                self.a[0].value == self.a[1].value
+                pass
+
+        @vsc.randobj
+        class Child2:
+            def __init__(self):
+                self.x = vsc.rand_list_t(vsc.attr(Field('x', 10)))
+                for i in range(5):    
+                    self.x.append(vsc.attr(Field('x', 10)))
+
+                self.y = vsc.rand_list_t(vsc.attr(Field('y', 10)))
+                for i in range(5):    
+                    self.y.append(vsc.attr(Field('y', 10)))
+    
+            @vsc.constraint
+            def test_c(self):
+                self.x[0].value < self.x[1].value
+                pass
+
+        inst=Parent()
+        inst.randomize(debug=0)
+        self.assertEqual(
+            inst.c1[0].a[1].value, 
+            inst.c2[0].x[1].value)
+            
+        print(inst.c1[0].a[0].value)
+        print(inst.c1[0].a[1].value)
+        print(inst.c2[0].x[0].value)
+        
+    def test_two_layer_2(self):
+        import vsc 
+
+        @vsc.randobj
+        class Parent:
+            def __init__(self):
+                self.id = 0
+                self.c1 = vsc.rand_list_t(vsc.rand_attr(Child1()))
+                for i in range(10):    
+                    self.c1.append(vsc.rand_attr(Child1()))      
+
+                self.c2 = vsc.rand_list_t(vsc.rand_attr(Child2()))
+                for i in range(10):
+                    self.c2.append(vsc.rand_attr(Child2()))
+
+            @vsc.constraint
+            def parent_c(self):
+                pass
+#                self.c1[0].a[1].value == self.c2[0].x[1].value       # Multi-level 
+                self.c1[0].a[1].value == self.c2[0].x[1].value
+
+#                with vsc.foreach(self.c1, idx=True) as i:
+#                    self.c1[i].a[0].value == self.c2[i].x[0].value
+
+
+        @vsc.randobj
+        class Field():
+            def __init__(self, name, def_value):
+                self.name = name
+                self.value = vsc.rand_uint8_t(def_value)
+
+        @vsc.randobj
+        class Child1:
+            def __init__(self):
+                self.a = vsc.rand_list_t(vsc.rand_attr(Field('an', 10)))
+                for i in range(5):    
+                    self.a.append(vsc.rand_attr(Field('an', 10)))
+
+                self.b = vsc.rand_list_t(vsc.rand_attr(Field('bn', 10)))
+                for i in range(5):    
+                    self.b.append(vsc.rand_attr(Field('bn', 10)))
+    
+            @vsc.constraint
+            def test_c(self):
+#                self.a[0].value < self.a[1].value
+#                self.a[0].value == self.a[1].value
+                pass
+
+        @vsc.randobj
+        class Child2:
+            def __init__(self):
+                self.x = vsc.rand_list_t(vsc.attr(Field('x', 10)))
+                for i in range(5):    
+                    self.x.append(vsc.attr(Field('x', 10)))
+
+                self.y = vsc.rand_list_t(vsc.attr(Field('y', 10)))
+                for i in range(5):    
+                    self.y.append(vsc.attr(Field('y', 10)))
+    
+            @vsc.constraint
+            def test_c(self):
+#                self.x[0].value < self.x[1].value
+                pass
+
+        inst=Parent()
+        inst.randomize(debug=0)
+        self.assertEqual(inst.c1[0].a[1].value, inst.c2[0].x[1].value)
+            
+        print(inst.c1[0].a[0].value)
+        print(inst.c1[0].a[1].value)
+        print(inst.c2[0].x[0].value)        
+        
+    def test_array_inst_constraint_mode(self):
+        import vsc 
+        
+        @vsc.randobj
+        class Parent:
+            def __init__(self):
+                self.id = 0
+                self.c1 = vsc.rand_list_t(vsc.rand_attr(Child1()))
+                for i in range(3):    
+                    self.c1.append(vsc.rand_attr(Child1()))
+
+        @vsc.randobj
+        class Child1():
+            def __init__(self):
+                self.a = vsc.rand_uint8_t()
+                self.b = vsc.rand_uint8_t()
+    
+            @vsc.constraint
+            def test_c(self):
+                self.a < 6
+                self.b < 3
+                self.a*self.b == 10
+
+        inst=Parent()
+        inst.c1[0].test_c.constraint_mode(False)
+        with inst.randomize_with():
+            inst.c1[0].a == 10
+            inst.c1[0].b == 20
+            pass
+
+        self.assertEqual(inst.c1[0].a, 10)
+        self.assertEqual(inst.c1[0].b, 20)
+        self.assertLess(inst.c1[1].a, 6)
+        self.assertLess(inst.c1[1].b, 3)
+        print("inst.c1[0].a =", inst.c1[0].a)
+        print("inst.c1[0].b =", inst.c1[0].b)
+        print("inst.c1[2].a =", inst.c1[2].a)
+        print("inst.c1[2].b =", inst.c1[2].b)        
+        
+